More size_t related issues have been found (thanks to the new snowpack server) and fixed. Basically, please remember that any variable used with/for the return value of a call to size() MUST be of type size_t (this becomes specially relevant in 64 bits).
The water triple point pressure has been updated to a better value.

+++ meteoio/A3DIO.h
@@ -48,7 +48,7 @@ class A3DIO : public IOInterface {
 
 		virtual void readMeteoData(const Date& dateStart, const Date& dateEnd,
 			                   std::vector< std::vector<MeteoData> >& vecMeteo,
-			                   const unsigned int& stationindex=IOUtils::npos);
+			                   const size_t& stationindex=IOUtils::npos);
 
 		virtual void writeMeteoData(const std::vector< std::vector<MeteoData> >& vecMeteo,
 		                            const std::string& name="");
