+++ meteoio/plugins/CosmoXMLIO.h
@@ -33,10 +33,10 @@ along with MeteoIO.  If not, see <http://www.gnu.org/licenses/>.
 #include <iostream>
 
 namespace mio {
-	
+
 	/**
 	* @class CosmoXMLIO
-	* @brief Reading of FieldExtra XML meteorological data. 
+	* @brief Reading of FieldExtra XML meteorological data.
 	* This is the plugin for reading the XML data genereated by FieldExtra, the post-processor
 	* of the MeteoSwiss COSMO meteorological model.
 	*
@@ -46,26 +46,26 @@ namespace mio {
 	class CosmoXMLIO : public IOInterface {
 		public:
 			CosmoXMLIO(void (*delObj)(void*), const Config& i_cfg);
-			
+
 			CosmoXMLIO(const std::string& configfile);
 			CosmoXMLIO(const CosmoXMLIO&);
 			CosmoXMLIO(const Config& cfgreader);
 			~CosmoXMLIO() throw();
-			
+
 			virtual void read2DGrid(Grid2DObject& grid_out, const std::string& parameter="");
-			
+
 			virtual void readDEM(DEMObject& dem_out);
 			virtual void readLanduse(Grid2DObject& landuse_out);
-			
+
 			virtual void readStationData(const Date& date, std::vector<StationData>& vecStation);
-			virtual void readMeteoData(const Date& dateStart, const Date& dateEnd, std::vector< std::vector<MeteoData> >& vecMeteo, const unsigned int& stationindex=IOUtils::npos);
-			
+			virtual void readMeteoData(const Date& dateStart, const Date& dateEnd, std::vector< std::vector<MeteoData> >& vecMeteo, const size_t& stationindex=IOUtils::npos);
+
 			virtual void writeMeteoData(const std::vector< std::vector<MeteoData> >& vecMeteo, const std::string& name="");
-			
+
 			virtual void readAssimilationData(const Date&, Grid2DObject& da_out);
 			virtual void readSpecialPoints(std::vector<Coords>& pts);
 			virtual void write2DGrid(const Grid2DObject& grid_in, const std::string& filename);
-			
+
 		private:
 			void cleanup() throw();
 			std::string getValue(xmlpp::TextReader& reader);
@@ -73,21 +73,21 @@ namespace mio {
 			Date getDateValue(xmlpp::TextReader& reader);
 			double c2k(const double& value);
 			double k2c(const double& value);
-			
-			void finishMeteo(const double& latitude, const double& longitude, const double& altitude, 
+
+			void finishMeteo(const double& latitude, const double& longitude, const double& altitude,
 			                 double& dew_point, MeteoData& meteo);
 			void writeHeader(std::stringstream& XMLdata);	//Write the first lines of the XML output file
 			void writeLocationHeader(const StationData& station, std::stringstream& XMLdata);
 			void writeMeteoDataDescription(std::stringstream& XMLdata);	//Write the middle of the XML output file
 			void writeMeteo(const std::vector<MeteoData>& vecMeteo, std::stringstream& XMLdata);
 			void writeFooter(std::stringstream& XMLdata);	//Write the last lines of the XML output file
-			
+
 			Config cfg;
 			double plugin_nodata;	//plugin specific no data value
 			static const double in_tz, out_tz; //plugin specific time zones
 			std::string coordin, coordinparam, coordout, coordoutparam; //projection parameters
 	};
-	
-	
+
+
 } //namespace
 #endif
